Chatbot_architecture/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-huggingface>=0.0.10
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
pypdf>=3.15.0
numpy>=1.24.0
pandas>=2.0.0